requirements.txt
fastapi
uvicorn
jinja2
requests
python-dotenv
sqlalchemy
pymysql
apscheduler
langchain_openai
langchain
langchain_community
pydef